+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Add global hooks ([#10](https://github.com/cucumber/javascript-core/pull/10))
 
 ## [0.3.0] - 2025-07-30
 ### Added